% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-package.R
\name{beta_CVX}
\alias{beta_CVX}
\alias{x}
\alias{y}
\title{Simulated data for the glmnet vignette}
\format{
Data objects used to demonstrate features in the glmnet vignette
}
\description{
Simple simulated data, used to demonstrate the features of glmnet
}
\details{
These datasets are artificial, and are used to test out some of the
features of glmnet.
}
\examples{

data(QuickStartExample)
x <- QuickStartExample$x; y <- QuickStartExample$y
glmnet(x, y)

}
\keyword{datasets}
