% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BinomialExample}
\alias{BinomialExample}
\title{Synthetic dataset with binary response}
\format{
List containing the following elements:
\describe{
\item{x}{100 by 30 matrix of numeric values.}
\item{y}{Numeric vector of length 100 containing 44 zeros and 56 ones.}
}
}
\usage{
data(BinomialExample)
}
\description{
Randomly generated data for binomial regression example.
}
\keyword{data}
