% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoxExample}
\alias{CoxExample}
\title{Synthetic dataset with right-censored survival response}
\format{
List containing the following elements:
\describe{
\item{x}{1,000 by 30 matrix of numeric values.}
\item{y}{1,000 by 2 matrix with column names "time" and "status". The
first column consists of positive numbers representing time to event,
while the second column represents the status indicator
(0=right-censored, 1=observed).}
}
}
\usage{
data(CoxExample)
}
\description{
Randomly generated data for Cox regression example.
}
\keyword{data}
