% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetFlex.R
\name{dev_function}
\alias{dev_function}
\title{Elastic net deviance value}
\usage{
dev_function(y, mu, weights, family)
}
\arguments{
\item{y}{Quantitative response variable.}

\item{mu}{Model's predictions for \code{y}.}

\item{weights}{Observation weights.}

\item{family}{A description of the error distribution and link function to be
used in the model. This is the result of a call to a family function.}
}
\description{
Returns the elastic net deviance value.
}
