% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-package.R
\name{glmnet-internal}
\alias{glmnet-internal}
\alias{auc}
\alias{assess.coxnet}
\alias{auc.mat}
\alias{cvtype}
\alias{cvstats}
\alias{cvcompute}
\alias{getcoef}
\alias{getcoef.multinomial}
\alias{fix.lam}
\alias{error.bars}
\alias{getmin}
\alias{elnet}
\alias{mrelnet}
\alias{lognet}
\alias{fishnet}
\alias{coefnorm}
\alias{cv.lognet}
\alias{cv.elnet}
\alias{cv.multnet}
\alias{cv.mrelnet}
\alias{cv.coxnet}
\alias{cv.fishnet}
\alias{cv.glmnet.raw}
\alias{cv.relaxed.raw}
\alias{blend.relaxed}
\alias{checkgamma.relax}
\alias{buildPredmat}
\alias{buildPredmat.mrelnetlist}
\alias{buildPredmat.multnetlist}
\alias{buildPredmat.lognetlist}
\alias{buildPredmat.array}
\alias{buildPredmat.coxnetlist}
\alias{buildPredmat.default}
\alias{lambda.interp}
\alias{nonzeroCoef}
\alias{glmnet_softmax}
\alias{getOptcv.glmnet}
\alias{getOptcv.relaxed}
\alias{jerr}
\alias{jerr.elnet}
\alias{jerr.lognet}
\alias{jerr.fishnet}
\alias{jerr.coxnet}
\alias{jerr.mrelnet}
\alias{plotCoef}
\alias{zeromat}
\alias{na.mean}
\alias{check_dots}
\alias{na_sparse_fix}
\alias{prepareX}
\title{Internal glmnet functions}
\description{
These are not intended for use by users. \code{lambda.interp} does linear
interpolation of the lambdas to obtain a prediction at a new point s.
\code{glmnet_softmax} does the classification for multinomial models.
\code{nonzeroCoef} determines in an efficient manner which variables are
nonzero in each fit. \code{jerr} prints out error messages from the C++ routines.
\code{plotCoef} is called by the \code{plot} method for \code{glmnet}
objects. \code{check_dots} is used in \code{coef} and \code{predict} with
argument \code{exact=TRUE}, to make sure user supplies original data used to
fit the \code{"glmnet"} object.
}
\author{
Trevor Hastie
}
\keyword{internal}
