% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.measures.R
\name{glmnet.measures}
\alias{glmnet.measures}
\title{Display the names of the measures used in CV for different "glmnet" families}
\usage{
glmnet.measures(
  family = c("all", "gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian",
    "GLM")
)
}
\arguments{
\item{family}{If a "glmnet" family is supplied, a list of the names of
measures available for that family are produced. Default is "all", in which
case the names of measures for all families are produced.}
}
\description{
Produces a list of names of measures
}
\details{
Try it and see. A very simple function to provide information
}
\seealso{
\code{cv.glmnet} and \code{assess.glmnet}.
}
\author{
Trevor Hastie\cr Maintainer: Trevor Hastie
\email{hastie@stanford.edu}
}
\keyword{models}
