% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MultiGaussianExample}
\alias{MultiGaussianExample}
\title{Synthetic dataset with multiple Gaussian responses}
\format{
List containing the following elements:
\describe{
\item{x}{100 by 20 matrix of numeric values.}
\item{y}{100 by 4 matrix of numeric values, each column representing
one response vector.}
}
}
\usage{
data(MultiGaussianExample)
}
\description{
Randomly generated data for multi-response Gaussian regression example.
}
\keyword{data}
