% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MultinomialExample}
\alias{MultinomialExample}
\title{Synthetic dataset with multinomial response}
\format{
List containing the following elements:
\describe{
\item{x}{500 by 30 matrix of numeric values.}
\item{y}{Numeric vector of length 500 containing 142 ones, 174 twos
and 184 threes.}
}
}
\usage{
data(MultinomialExample)
}
\description{
Randomly generated data for multinomial regression example.
}
\keyword{data}
