% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.coxnet.R
\name{mycoxph}
\alias{mycoxph}
\title{Helper function to fit coxph model for survfit.coxnet}
\usage{
mycoxph(object, s, ...)
}
\arguments{
\item{object}{A class \code{coxnet} object.}

\item{s}{The value of the penalty parameter lambda at which the survival
curve is required.}

\item{...}{The same ... that was passed to survfit.coxnet.}
}
\description{
This function constructs the coxph call needed to run the "hack" of
coxph with 0 iterations. It's a separate function as we have to deal with
function options like strata, offset and observation weights.
}
