% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.glmnet.R
\name{objective.glmnet}
\alias{objective.glmnet}
\title{Penalized log-likelihood for a glmnet object}
\usage{
objective.glmnet(object)
}
\arguments{
\item{object}{fitted \code{glmnet} object}
}
\value{
Numeric vector of penalized log-likelihood values.
}
\description{
Compute the penalized log-likelihood sequence from the glmnet object.
}
\examples{
 x = matrix(rnorm(100 * 20), 100, 20)
 y = rnorm(100)
 fit = glmnet(x, y)
 objective.glmnet(fit)
}
