% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PoissonExample}
\alias{PoissonExample}
\title{Synthetic dataset with count response}
\format{
List containing the following elements:
\describe{
\item{x}{500 by 20 matrix of numeric values.}
\item{y}{Numeric vector of length 500 consisting of non-negative
integers.}
}
}
\usage{
data(PoissonExample)
}
\description{
Randomly generated data for Poisson regression example.
}
\keyword{data}
