% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QuickStartExample}
\alias{QuickStartExample}
\title{Synthetic dataset with Gaussian response}
\format{
List containing the following elements:
\describe{
\item{x}{100 by 20 matrix of numeric values.}
\item{y}{Numeric vector of length 100.}
}
}
\usage{
data(QuickStartExample)
}
\description{
Randomly generated data for Gaussian regression example.
}
\keyword{data}
