% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.coxnet.R
\name{response.coxnet}
\alias{response.coxnet}
\title{Make response for coxnet}
\usage{
response.coxnet(y)
}
\arguments{
\item{y}{Response variable. Either a class "Surv" object or a two-column
matrix with columns named 'time' and 'status'.}
}
\value{
A class "Surv" object.
}
\description{
Internal function to make the response y passed to glmnet suitable
for coxnet (i.e. glmnet with family = "cox"). Sanity checks are performed
here too.
}
\details{
If y is a class "Surv" object, this function returns y with no changes. If
y is a two-column matrix with columns named 'time' and 'status', it is
converted into a "Surv" object.
}
