% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmult.R
\name{rmult}
\alias{rmult}
\title{Generate multinomial samples from a probability matrix}
\usage{
rmult(p)
}
\arguments{
\item{p}{matrix of probabilities, with number of columns the number of classes}
}
\value{
a vector of class memberships
}
\description{
Generate multinomial samples
}
\details{
Simple function that calls the \code{rmultinom} function. It generates a class label
for each row of its input matrix of class probabilities.
}
\author{
Trevor Hastie \cr Maintainer: Trevor Hastie \href{mailto:hastie@stanford.edu}{hastie@stanford.edu}
}
