% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SparseExample}
\alias{SparseExample}
\title{Synthetic dataset with sparse design matrix}
\format{
List containing the following elements:
\describe{
\item{x}{100 by 20 matrix of numeric values. x is in sparse matrix
format, having class "dgCMatrix".}
\item{y}{Numeric vector of length 100.}
}
}
\usage{
data(SparseExample)
}
\description{
Randomly generated data for Gaussian regression example with the
design matrix x being in sparse matrix format.
}
\keyword{data}
