% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpath.R
\name{use.cox.path}
\alias{use.cox.path}
\title{Check if glmnet should call cox.path}
\usage{
use.cox.path(x, y)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Response variable.}
}
\value{
TRUE if cox.path() should be called, FALSE otherwise.
}
\description{
Helper function to check if glmnet() should call cox.path().
}
\details{
For \code{family="cox"}, we only call the original coxnet() function if
(i) x is not sparse, (ii) y is right-censored data, and (iii) we are
not fitting a stratified Cox model. This function also throws an error
if y has a "strata" attribute but is not of type "stratifySurv".
}
